requirements.txt
python-dotenv
pandas
numpy
pymongo==4.10.1
certifi
pymongo[srv]==3.6

#-e .